+++ src/spatialdata_io/__init__.py
@@ -1,6 +1,5 @@
 from importlib.metadata import version
 
-from spatialdata_io.readers._utils._utils import xenium_explorer_selection
 from spatialdata_io.readers.codex import codex
 from spatialdata_io.readers.cosmx import cosmx
 from spatialdata_io.readers.curio import curio
@@ -8,7 +7,7 @@ from spatialdata_io.readers.mcmicro import mcmicro
 from spatialdata_io.readers.merscope import merscope
 from spatialdata_io.readers.steinbock import steinbock
 from spatialdata_io.readers.visium import visium
-from spatialdata_io.readers.xenium import xenium
+from spatialdata_io.readers.xenium import xenium, xenium_aligned_image, xenium_explorer_selection
 
 __all__ = [
     "curio",
@@ -19,6 +18,7 @@ __all__ = [
     "mcmicro",
     "steinbock",
     "merscope",
+    "xenium_aligned_image",
     "xenium_explorer_selection",
 ]
