+++ src/spatialdata_io/readers/_utils/_utils.py
@@ -69,25 +69,3 @@ def _read_counts(
     adata.uns["spatial"] = {library_id: {"metadata": {}}}  # can overwrite
     return adata, library_id
 
-
-def xenium_explorer_selection(path: str | Path, pixelsize: float = 0.2125) -> Polygon:
-    """Read the coordinates of a selection file exported from the Xenium Explorer.
-
-    This `.csv` file can be generated by the "Freehand Selection" as well
-    as the "Rectangular Selection".
-    The output `Polygon` can be used for a polygon query on the pixel coordinate
-    system (called `"global"` for Xenium data).
-
-    Parameters
-    ----------
-    path
-        Path to the `.csv` file containing the selection coordinates
-    pixelsize
-        Size of a pixel in microns. By default, the Xenium pixel size is used.
-
-    Returns
-    -------
-    :class:`shapely.geometry.Polygon`
-    """
-    df = pd.read_csv(path, skiprows=2)
-    return Polygon(df.values / pixelsize)
