+++ src/spatialdata_io/readers/xenium.py
@@ -28,7 +28,7 @@ from spatialdata_io._docs import inject_docs
 from spatialdata_io._utils import deprecation_alias
 from spatialdata_io.readers._utils._read_10x_h5 import _read_10x_h5
 
-__all__ = ["xenium"]
+__all__ = ["xenium", "xenium_aligned_image", "xenium_explorer_selection"]
 
 
 @deprecation_alias(cells_as_shapes="cells_as_circles")
@@ -325,3 +325,27 @@ def xenium_aligned_image(
         transformations={"global": transformation},
         **image_models_kwargs,
     )
+
+
+def xenium_explorer_selection(path: str | Path, pixelsize: float = 0.2125) -> Polygon:
+    """Read the coordinates of a selection file exported from the Xenium Explorer.
+
+    This `.csv` file can be generated by the "Freehand Selection" as well
+    as the "Rectangular Selection".
+    The output `Polygon` can be used for a polygon query on the pixel coordinate
+    system (called `"global"` for Xenium data).
+
+    Parameters
+    ----------
+    path
+        Path to the `.csv` file containing the selection coordinates
+    pixelsize
+        Size of a pixel in microns. By default, the Xenium pixel size is used.
+
+    Returns
+    -------
+    :class:`shapely.geometry.Polygon`
+    """
+    df = pd.read_csv(path, skiprows=2)
+    return Polygon(df.values / pixelsize)
+
